Optionally extract quiz ID from a custom field
This code change allows the quiz ID to be specified from a custom field from a post.

This has allowed me to put Slickquiz in a sidebar which calls different quizzes depending on the page. (for my purposes I've fixed it in the sidebar so comprehension questions are always visible on the side when scrolling through an article).

Note: this is my first ever attempt at coding something for 'real' purposes or using github to make changes!